requirements.txt
pytesseract>=0.3.10
opencv-python>=4.5.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
